+++ packages/_vue3-migration-test/src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './animations';
 export { default as TestBaseDropdown } from './test-base-dropdown.vue';
 export * from './scroll';
+export { default as TestBaseEventButton } from './test-base-event-button.vue';