+++ packages/x-components/src/components/animations/fade.vue
@@ -19,9 +19,7 @@
     name: 'Fade',
     inheritAttrs: false,
     props: {
-      /**
-       * Indicates if the transition must be applied on the initial render of the node.
-       */
+      /** Indicates if the transition must be applied on the initial render of the node. */
       appear: {
         type: Boolean,
         default: true