+++ packages/x-components/src/x-modules/search/components/__tests__/sort-picker-list.spec.ts
@@ -92,7 +92,7 @@ describe('testing SortPickerList component', () => {
       metadata: {
         moduleName: 'search',
         target: getButton(2),
-        location: undefined, // TODO - Emission by BaseEventButton in the old way. `none` when use$x
+        location: 'none',
         replaceable: true
       }
     });